fix: screenshot background processing
