+++ PostHog.xcodeproj/xcshareddata/xcschemes/PostHogTests.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1530"
+   LastUpgradeVersion = "1540"
    version = "2.2">
    <BuildAction
       parallelizeBuildables = "YES"
