+++ PostHog/Replay/RRWireframe.swift
@@ -8,6 +8,7 @@
 //
 
 import Foundation
+import UIKit
 
 class RRWireframe {
     var id: Int = 0
@@ -21,6 +22,8 @@ class RRWireframe {
     var text: String?
     var label: String?
     var value: Any? // string or number
+    var image: UIImage? // clean reference after using it
+    var maskableWidgets: [CGRect]?
     var base64: String?
     var style: RRStyle?
     var disabled: Bool?
@@ -29,6 +32,36 @@ class RRWireframe {
     var max: Int?
     // internal
     var parentId: Int?
+    
+    private func imageToBase64(_ image: UIImage) -> String? {
+        let jpegData = image.jpegData(compressionQuality: 0.3)
+        let base64 = jpegData?.base64EncodedString()
+
+        if let base64 = base64 {
+            return "data:image/jpeg;base64,\(base64)"
+        }
+
+        return nil
+    }
+    
+    private func maskImage() -> UIImage? {
+        if let image = image {
+            if let maskableWidgets = maskableWidgets {
+                let redactedImage = UIGraphicsImageRenderer(size: image.size, format: .init(for: .init(displayScale: 1))).image { context in
+                    context.cgContext.interpolationQuality = .none
+                    image.draw(at: .zero)
+
+                    for rect in maskableWidgets {
+                        let path = UIBezierPath(roundedRect: rect, cornerRadius: 10)
+                        UIColor.black.setFill()
+                        path.fill()
+                    }
+                }
+                return redactedImage
+            }
+        }
+        return nil
+    }
 
     func toDict() -> [String: Any] {
         var dict: [String: Any] = [
@@ -62,6 +95,15 @@ class RRWireframe {
         if let value = value {
             dict["value"] = value
         }
+        
+        if let image = image {
+            if let maskedImage = maskImage() {
+                base64 = imageToBase64(maskedImage)
+            } else {
+                base64 = imageToBase64(image)
+            }
+        }
+        image = nil
 
         if let base64 = base64 {
             dict["base64"] = base64
